+++ app/src/app/shared/riwayat/penguasaan-bahasa/penguasaan-bahasa.component.ts
@@ -14,6 +14,7 @@ import { Subject, takeUntil } from 'rxjs';
 import { Observable } from 'rxjs/internal/Observable';
 import { CreateComponent } from 'app/modules/referensi/bahasa/create/create.component';
 import { MatDialog } from '@angular/material/dialog';
+import { FuseConfirmationService } from '@fuse/services/confirmation';
 
 @Component({
     selector: 'app-penguasaan-bahasa',
@@ -42,6 +43,7 @@ export class PenguasaanBahasaComponent implements OnInit, OnDestroy {
         private _referensiService: ReferensiService,
         private _helperService: HelperService,
         private _changeDetectorRef: ChangeDetectorRef,
+        private _fuseConfirmationService: FuseConfirmationService,
         private _formBuilder: FormBuilder,
         private _toastr: ToastrService,
         private _authService: AuthService,
@@ -139,18 +141,44 @@ export class PenguasaanBahasaComponent implements OnInit, OnDestroy {
     }
 
     delete(): void {
-        this._penerjemahService.deleteKemahiranBahasa(this.selected.pnsId, this.selected.bahasa.id).subscribe(
-            (result) => {
-                if (result?.success) {
-                    this._toastr.success('Selanjutnya usulan Anda akan diproses oleh Admin', 'Usulan Hapus Berhasil');
-                    this.loadData();
-                    this.toggleInsertMode(false);
-                    this.toggleEditMode(false);
-                } else {
-                    this._toastr.error(result?.message, 'ERROR');
+        const dialogRef = this._fuseConfirmationService.open({
+            'title': 'Hapus Data',
+            'message': '<span class="font-medium">Tindakan ini tidak dapat dibatalkan!</span>',
+            'icon': {
+                'show': true,
+                'name': 'heroicons_outline:x',
+                'color': 'warn'
+            },
+            'actions': {
+                'confirm': {
+                    'show': true,
+                    'label': 'Konfirm hapus',
+                    'color': 'warn'
+                },
+                'cancel': {
+                    'show': true,
+                    'label': 'Batal'
                 }
+            },
+            'dismissible': true
+        });
+
+        dialogRef.afterClosed().subscribe((_result) => {
+            if (_result === 'confirmed') {
+                this._penerjemahService.deleteKemahiranBahasa(this.selected.pnsId, this.selected.bahasa.id).subscribe(
+                    (result) => {
+                        if (result?.success) {
+                            this._toastr.success('Selanjutnya usulan Anda akan diproses oleh Admin', 'Usulan Hapus Berhasil');
+                            this.loadData();
+                            this.toggleInsertMode(false);
+                            this.toggleEditMode(false);
+                        } else {
+                            this._toastr.error(result?.message, 'ERROR');
+                        }
+                    }
+                );
             }
-        );
+        });
     }
 
     update(): void {
