+++ app/src/app/shared/riwayat/uji-kompetensi/uji-kompetensi.component.ts
@@ -16,6 +16,7 @@ import { CertificateHistoryService } from 'app/modules/survey-kuisoner/sertifika
 import { PdfGenerationService } from 'app/services/pdf-generation.service';
 import { ViewChild, ElementRef } from '@angular/core';
 import { ToastrService } from 'ngx-toastr';
+import { FuseConfirmationService } from '@fuse/services/confirmation';
 
 @Component({
     selector: 'app-uji-kompetensi',
@@ -51,6 +52,7 @@ export class UjiKompetensiComponent implements OnInit, OnDestroy {
         private _referensiService: ReferensiService,
         private _helperService: HelperService,
         private _changeDetectorRef: ChangeDetectorRef,
+        private _fuseConfirmationService: FuseConfirmationService,
         private _formBuilder: FormBuilder,
         private _toastr: ToastrService,
         private _authService: AuthService,
@@ -295,18 +297,44 @@ export class UjiKompetensiComponent implements OnInit, OnDestroy {
     }
 
     delete(): void {
-        this._penerjemahService.deleteRwUjiKompetensi(this.selected.id).subscribe(
-            (result) => {
-                if (result?.success) {
-                    this._toastr.success('Selanjutnya usulan Anda akan diproses oleh Admin', 'Usulan Hapus Berhasil');
-                    this.loadData();
-                    this.toggleInsertMode(false);
-                    this.toggleEditMode(false);
-                } else {
-                    this._toastr.error(result?.message, 'ERROR');
+        const dialogRef = this._fuseConfirmationService.open({
+            'title': 'Hapus Data',
+            'message': '<span class="font-medium">Tindakan ini tidak dapat dibatalkan!</span>',
+            'icon': {
+                'show': true,
+                'name': 'heroicons_outline:x',
+                'color': 'warn'
+            },
+            'actions': {
+                'confirm': {
+                    'show': true,
+                    'label': 'Konfirm hapus',
+                    'color': 'warn'
+                },
+                'cancel': {
+                    'show': true,
+                    'label': 'Batal'
                 }
+            },
+            'dismissible': true
+        });
+
+        dialogRef.afterClosed().subscribe((_result) => {
+            if (_result === 'confirmed') {
+                this._penerjemahService.deleteRwUjiKompetensi(this.selected.id).subscribe(
+                    (result) => {
+                        if (result?.success) {
+                            this._toastr.success('Selanjutnya usulan Anda akan diproses oleh Admin', 'Usulan Hapus Berhasil');
+                            this.loadData();
+                            this.toggleInsertMode(false);
+                            this.toggleEditMode(false);
+                        } else {
+                            this._toastr.error(result?.message, 'ERROR');
+                        }
+                    }
+                );
             }
-        );
+        });
     }
 
     update(): void {
